feat: allow enrollment api admin to see all enrollments
Co-authored-by: Maxim Beder <maxim@opencraft.com>

test: EnterpriseCourseEnrollmentFilterBackend

Co-authored-by: Maxim Beder <maxim@opencraft.com>

feat: add additional fields to EnterpriseCourseEnrollmentViewSet

Created new manager and serializer for EnterpriseCourseEnrollment model,
which add additional fields to the model, which allow to reduce the
number of requests to the API by providing the necessary data in the
feilds along side the model.

Use the new manager and serializer in EnterpriseCourseEnrollmentViewSet.

Co-authored-by: Maxim Beder <maxim@opencraft.com>