feat: create CourseEnrollmentAllowed entries for pending enrollments
When creating pending enrollments for non-existant users, we also check
to see if the course is "invite_only". If the course is invite only,
then we create corresponding CourseEnrollmentAllowed objects. This fixes
the issue when the enterprise creates pending enrollment, but the user
cannot enroll to the course as platform rejects the enrollment request
due to missing CEA for the user.

(cherry picked from commit 0e06d7425488a390c6a49fd0e877a28f71437256)